requirements.txt
scikit-learn
joblib
boto3
python-dotenv